requirements.txt
alembic==1.16.1
bleach==6.2.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dotenv==0.9.9
Flask==2.3.3
Flask-Dance==7.1.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==3.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
oauthlib==3.2.2
python-dotenv==1.1.0
requests==2.32.3
requests-oauthlib==2.0.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
urllib3==2.4.0
URLObject==2.4.3
webencodings==0.5.1
Werkzeug==2.3.7
WTForms==3.0.1
qrcode[pil]
faiss-cpu==1.11.0
sentence-transformers


